// e-livraria-web/src/models/EOrderStatus.ts
export enum EOrderStatus {
    EM_PROCESSAMENTO = "EM_PROCESSAMENTO",
    APROVADA = "APROVADA",
    REPROVADA = "REPROVADA",
    EM_TRANSITO = "EM_TRANSITO",
    ENTREGUE = "ENTREGUE",
    EM_TROCA = "EM_TROCA",
    TROCA_AUTORIZADA = "TROCA_AUTORIZADA",
    TROCA_NEGADA = "TROCA_NEGADA",
    EM_TRANSPORTE = "EM_TRANSPORTE",
    TROCADO = "TROCADO",
    CANCELADO = "CANCELADO",
    EM_DEVOLUCAO = "EM_DEVOLUCAO",
    DEVOLVIDO = "DEVOLVIDO"
}